Got Teleop automation working, can now score a cone with three button presses. Started work on delivering a cone in auto. Fixed an error in the SignalDetectionPipeline. Removed teleop state logic in NewTeleopBase.

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Commands/ftclib/TelemetryCommand.java
@@ -11,7 +11,7 @@ public class TelemetryCommand extends CommandBase {
     private final String msg;
     private final WaitCommand waitCommand;
 
-    public TelemetryCommand(Telemetry telemetry, String msg, long duration) {
+    public TelemetryCommand(Telemetry telemetry, String msg, double duration) {
         this.telemetry = telemetry;
         this.msg = msg;
         waitCommand = new WaitCommand((long) (duration * 1000));
